Saving of an article works but removes the category https://github.com/WomenPlusPlus/deploy-impact-22-wikimini-b/issues/188
