+++ src/node-backend/controller/homework-controller.js
@@ -51,47 +51,9 @@ export async function updateArticleTitleForHwTask(articleTitle, articleUrl) {
 
 }
 
-// works if the selectTasksTemplateQuery is used
-function convertDbTasks(dbResult) {
-    const tasks = [];
-    for (let i = 0; i < dbResult.length; i++) {
-        const taskId = dbResult[i]['taskId'];
-        const title = dbResult[i]['title'];
-        const description = dbResult[i]['description'];
-        const assignedByTeacher = dbResult[i]['assignedByTeacher'];
-        const assignedToStudent = dbResult[i]['assignedToStudent'];
-        const startDate = dbResult[i]['startdate'];
-        const dueDate = dbResult[i]['duedate'];
-        const hwType = Object.keys(HwType).find(k => HwType[k]['id'] === dbResult[i]['type']);
-        const doneDate = dbResult[i]['donedate'] === null ? false : dbResult[i]['donedate'];
-        const gradedDate = dbResult[i]['gradeddate'] === null ? false : dbResult[i]['gradeddate'];
-        const status = Object.keys(Status).find(k => Status[k]['id'] === dbResult[i]['status']);
-        const article = dbResult[i]['concernsArticle'] === null ? false : dbResult[i]['concernsArticle'];
-        const gradingCategories = [];
-        while (taskId === dbResult[i]['taskId']) {
-            gradingCategories[i] = dbResult[i]['categoryId'];
-            i++;
-            if (i === dbResult.length - 1) {
-                break;
-            }
-        }
-        const task = new HwTask(title, description,
-            assignedByTeacher, assignedToStudent,
-            gradingCategories, startDate, dueDate);
-        task.id = taskId;
-        task.hwType = hwType;
-        task.doneDate = doneDate;
-        task.gradedDate = gradedDate;
-        task.status = status;
-        task.articleTitle = article;
-        tasks[tasks.length] = task;
-    }
-    return tasks;
-}
-
 export async function getHwTask(hwTaskId) {
     const result = await dbAdapter.getHwTask(hwTaskId);
-    return convertDbTasks(result)[0];
+    return dbAdapter.convertDbTasks(result)[0];
 }
 
 function isCurrent(task) {
@@ -112,7 +74,7 @@ function isGraded(task) {
 
 export async function getHwForStudent(username) {
     const result = await dbAdapter.getHwForStudent(username);
-    const allHomework = convertDbTasks(result);
+    const allHomework = dbAdapter.convertDbTasks(result);
     const current = allHomework.filter(task => isCurrent(task));
     const late = allHomework.filter(task => isLate(task));
     const done = allHomework.filter(task => isDone(task));
